Assignments points 1,2,3,5 and 6 completed

+++ app/src/main/java/com/mpg/nagarro/deviceinventorymgmt/ui/devices/DeviceListViewModel.kt
@@ -1,6 +1,5 @@
 package com.mpg.nagarro.deviceinventorymgmt.ui.devices
 
-import android.util.Log
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.viewModelScope
 import com.mpg.nagarro.deviceinventorymgmt.base.BaseViewModel
@@ -12,14 +11,18 @@ import javax.inject.Inject
 class DeviceListViewModel @Inject constructor(private val repository: Repository) :
     BaseViewModel() {
 
-    private val TAG = "DeviceListViewModel"
-
     val devices: LiveData<List<DeviceEntity>> = repository.getDeviceList()
 
-    fun deleteRow(item: DeviceEntity) = viewModelScope.launch {
-        val position = devices.value?.indexOf(item)
-        Log.i(TAG, "$position Row : ${item.name}")
-        repository.deleteDevice(item.deviceId)
+    fun deleteRowAction(item: DeviceEntity) = viewModelScope.launch {
+
+        val deviceEntity = repository.getDeviceById(item.deviceId)
+        deviceEntity?.let {
+            if (it.totalInventory == it.currentAvailableInventory)
+                repository.deleteDevice(item.deviceId)
+            else
+                showMessage.postValue("\" ${item.name} \" can't be deleted \n  Inventory either not returned or lost. ")
+        }
+
     }
 
 }
